Forced the use of page/post.description
Both in meta tags and in the postcard. This means that the description
can be removed from the main body of the text and put into the front
matter. Consequently, the CSS rule which increased the size of the first
paragraph of the blog post has been removed as well.