HRM Post v0.9
BIG commit containing needed pictures, the KLEs that were used for the
pictures, changes to the CSS and layout HTML files to make the home row
mods post look like how I want it to look like.

The raw prose content of the home row mods post is mostly complete. All
that's left is to write the section on alternative home row mods layouts
and little notes and corrections here and there.