Removed some scrollbar CSS to fix(?) double scrlbr
Sometimes, when loading the home row mods page in particular, you get
two scrollbars instead of just one. The extra scollbar only being
capable of a limited range of motion and any content outside of that
limited range is scrambled and jammed all on top of each other. A simple
refresh or Ctrl+Shift+I is enough to fix the issue. I have yet to find a
way to reliably reproduce the problem.